Divide `sponsor.py` em `sponsor.py` e `std_sponsor.py` para evitar dependencia circular
